+++ vector/src/main/java/im/vector/activity/WidgetActivity.kt
@@ -18,13 +18,13 @@
 package im.vector.activity
 
 import android.annotation.SuppressLint
-import android.app.AlertDialog
 import android.content.Context
 import android.content.Intent
 import android.graphics.Bitmap
 import android.os.Build
 import android.os.Handler
 import android.os.Looper
+import android.support.v7.app.AlertDialog
 import android.text.TextUtils
 import android.view.View
 import android.view.ViewGroup