+++ vector/src/main/java/im/vector/dialogs/ConsentNotGivenHelper.kt
@@ -17,8 +17,8 @@
 package im.vector.dialogs
 
 import android.app.Activity
-import android.app.AlertDialog
 import android.os.Bundle
+import android.support.v7.app.AlertDialog
 import im.vector.Matrix
 import im.vector.R
 import im.vector.activity.SimpleWebViewActivity