+++ vector/src/main/java/im/vector/util/RageShake.java
@@ -17,7 +17,6 @@
 
 package im.vector.util;
 
-import android.app.AlertDialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.hardware.Sensor;
@@ -25,6 +24,7 @@ import android.hardware.SensorEvent;
 import android.hardware.SensorEventListener;
 import android.hardware.SensorManager;
 import android.os.Build;
+import android.support.v7.app.AlertDialog;
 
 import org.matrix.androidsdk.util.Log;