+++ vector/src/main/java/im/vector/util/SlashComandsParser.java
@@ -17,7 +17,7 @@
 
 package im.vector.util;
 
-import android.app.AlertDialog;
+import android.support.v7.app.AlertDialog;
 import android.text.TextUtils;
 import android.widget.Toast;